+++ gaussdb/include/huaweicloud/gaussdb/v3/model/ScalingStrategyReqInfo.h
@@ -48,7 +48,7 @@ public:
     void setFlavorSwitch(const std::string& value);
 
     /// <summary>
-    /// 增删只读节点开关。自动增删只读节点功能需要先[开启数据库代理](https://support.huaweicloud.com/api-gaussdbformysql/CreateGaussMySqlProxy.html)，且只能有一个代理。 - ON：开启。 - OFF：关闭。
+    /// 增删只读节点开关。自动增删只读节点功能需要先[开启数据库代理](https://support.huaweicloud.com/api-taurusdb/CreateGaussMySqlProxy.html)，且只能有一个代理。 - ON：开启。 - OFF：关闭。
     /// </summary>
 
     std::string getReadOnlySwitch() const;